pyproject.toml
[project]
name = "knowledge"
version = "0.0.1"
description = "Database"
authors = [{ name = "Raphael Sourty", email = "raphael.sourty@gmail.com" }]
readme = "readme.md"
requires-python = ">=3.10,<3.13"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]

dependencies = [
    "beautifulsoup4==4.9.3",
    "pandas==1.5.3",
    "requests==2.32.2",
    "rdflib==6.1.1",
    "cherche==2.2.1",
    "networkx==2.8.5",
    "uvicorn==0.17.5",
    "fastapi==0.116.1",
    "pydantic==2.11.7",
    "python-multipart==0.0.18",
    "pyyaml_env_tag==0.1",
    "pyzotero==1.5.5",
    "scikit-learn==1.5.0",
    "openai==1.35.2",
    "orjson==3.9.15",
    "lenlp==1.1.0",
    "sentence_transformers==5.1.2",
    "numpy==2.0.0",
    "trafilatura==2.0.0",
    "transformers==4.56.2",
    "huggingface-hub==0.36.0",
    "tokenizers==0.22.1",
]

[project.optional-dependencies]
dev = ["ruff>=0.8.0", "mypy>=1.13.0", "pre-commit>=4.0.0"]

[project.urls]
Homepage = "https://github.com/raphaelsty/knowledge"
Repository = "https://github.com/raphaelsty/knowledge"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["knowledge_database", "api"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["knowledge_database"]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
check_untyped_defs = false
disable_error_code = [
    "assignment",
    "attr-defined",
    "arg-type",
    "var-annotated",
    "dict-item",
    "no-any-return",
    "import-untyped",
]
